requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-richtextfield==1.6.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.3.0
drf-friendly-errors-2==0.14.3
idna==2.10
mysqlclient==2.0.3
Pillow==8.2.0
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
uWSGI==2.0.20
